# swagger.yaml
swagger: '2.0'
info:
  title: API Documentation
  description: API documentation using drf-yasg
  contact:
    email: tangjzh23@mail2.sysu.edu.cn
  version: v1
basePath: /
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Basic:
    type: basic
security:
- Basic: []
paths:
...
  /social/post/:
    get:
      operationId: social_post_list
      description: 获取所有车小圈帖子
      parameters:
      - name: page
        in: query
        description: A page number within the paginated result set.
        required: false
        type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
            - count
            - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/Post'
      tags:
      - social
    post:
      operationId: social_post_create
      description: 创建车小圈帖子
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/Post'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/Post'
      tags:
      - social
    parameters: []
...
definitions:
...
  Tag:
    required:
    - name
    type: object
    properties:
      name:
        title: Name
        type: string
        maxLength: 100
        minLength: 1
      slug:
        title: Slug
        type: string
        format: slug
        pattern: ^[-a-zA-Z0-9_]+$
        maxLength: 100
  PostImage:
    type: object
    properties:
      image:
        title: Image
        type: string
        readOnly: true
        format: uri
  Post:
    type: object
    properties:
      id:
        title: Id
        type: string
        format: uuid
        readOnly: true
      user:
        title: User
        type: string
        readOnly: true
      title:
        title: Title
        type: string
        maxLength: 255
        minLength: 1
        x-nullable: true
      content:
        title: Content
        type: string
        minLength: 1
      created_at:
        title: Created at
        type: string
        format: date-time
        readOnly: true
      updated_at:
        title: Updated at
        type: string
        format: date-time
        readOnly: true
      tags:
        type: array
        items:
          $ref: '#/definitions/Tag'
      location:
        title: Location
        type: string
        maxLength: 255
        x-nullable: true
      likes_count:
        title: Likes count
        type: integer
        default: 0
      views_count:
        title: Views count
        type: integer
        default: 0
      images:
        type: array
        items:
          $ref: '#/definitions/PostImage'
      comments:
        type: array
        items:
          type: integer
        readOnly: true
        uniqueItems: true
...